move DEPARTEMENTS* static variables to dedicated module
