+++ labonneboite/common/util.py
@@ -18,18 +18,6 @@ logger = logging.getLogger('main')
 ENABLE_TIMEIT_TIMERS = True
 
 
-def get_departements(largest_ones_first=False):
-    departements = ["{:02d}".format(d) for d in range(1, 96)] + ['97']
-    if largest_ones_first:
-        departements.remove('75')
-        departements[:0] = ['75']
-    return departements
-
-
-DEPARTEMENTS = get_departements()
-DEPARTEMENTS_WITH_LARGEST_ONES_FIRST = get_departements(largest_ones_first=True)
-
-
 def timeit(func):
     @wraps(func)
     def wrap(*args, **kw):
