+++ labonneboite/tests/app/test_synchro.py
@@ -9,7 +9,7 @@ from sqlalchemy import create_engine
 from labonneboite.common.database import db_session, get_db_string
 from labonneboite.common.database import ENGINE_PARAMS, REAL_DATABASE
 from labonneboite.common.models import Office
-from labonneboite.common import util
+from labonneboite.common import departements as dpt
 
 
 class SynchronizationIndexAndDatabaseTest(unittest.TestCase):
@@ -47,7 +47,7 @@ class SynchronizationIndexAndDatabaseTest(unittest.TestCase):
         """
         Test that the data is synchronized between the SQL database and Elasticsearch.
         """
-        departement = random.choice(util.DEPARTEMENTS)
+        departement = random.choice(dpt.DEPARTEMENTS)
         offices = Office.query.filter(Office.departement == departement).limit(100)
 
         # If we got no results here, assume that tests are running locally,
